Move preset into src
