+++ packages/primitives/src/index.ts
@@ -11,3 +11,9 @@ export type { ButtonProps } from "./Button";
 export { IconButton } from "./IconButton";
 export type { IconButtonProps } from "./IconButton";
 export { Input, InputContainer, TextArea } from "./Input";
+export { Icon } from "./Icon";
+export type { IconProps } from "./Icon";
+export { Spinner } from "./Spinner";
+export type { SpinnerProps } from "./Spinner";
+
+export { preset } from "./preset";
